agent-python-backend/requirements.api.txt
fastapi
uvicorn
pandas
google-cloud-aiplatform
google-cloud-secret-manager
openai
httpx
python-multipart
pillow
redis
rq
matplotlib
uvloop
httptools
scipy
beautifulsoup4==4.12.3
lxml
playwright
lightweight_mmm==0.1.9
jax==0.4.23
jaxlib==0.4.23
numpyro==0.13.2
arviz==0.15.1